closes #5. download and present Option Positions

+++ db/migrations/V0005__create_option_groupings.sql
@@ -1,5 +1,5 @@
 CREATE TABLE option_groupings (
-    cid BIGINT PRIMARY KEY,
+    cid SERIAL PRIMARY KEY,
     name VARCHAR(255)
 );
